+++ catalog-service/src/main/java/com/node5/catalogservice/product/application/ProductService.java
@@ -1,10 +1,7 @@
 package com.node5.catalogservice.product.application;
 
 import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Map;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -126,48 +123,9 @@ public class ProductService {
 			.map(ProductInfo::from);
 	}
 
-	@Transactional(readOnly = true)
-	public Map<UUID, UUID> getShopIdsByProductIds(List<UUID> productIds) {
-
-		if (productIds == null || productIds.isEmpty()) {
-			return Map.of();
-		}
-
-		List<UUID> distinctIds = productIds.stream().distinct().toList();
-
-		List<Product> products = productRepository.findAllByIdIn(distinctIds);
-
-		if (products.size() != distinctIds.size()) {
-			throw new BaseException(ProductErrorCode.PRODUCT_NOT_FOUND);
-		}
-
-		return products.stream()
-			.collect(Collectors.toMap(Product::getId, Product::getShopId));
-	}
-
-	@Transactional(readOnly = true)
-	public List<Product> getProductsByIds(List<UUID> productIds) {
-		if (productIds == null || productIds.isEmpty()) {
-			return List.of();
-		}
-		return productRepository.findAllByIdInAndStatus(productIds, ProductStatus.ON_SALE);
-	}
-
 	private Product getProductOrThrow(UUID productId) {
 		return productRepository.findById(productId)
 			.orElseThrow(() -> new BaseException(ProductErrorCode.PRODUCT_NOT_FOUND));
-	}
-
-	@Transactional(readOnly = true)
-	public List<UUID> getOnSaleProductIds(Pageable pageable) {
-		return productRepository.findByStatus(ProductStatus.ON_SALE, pageable)
-			.map(Product::getId)
-			.getContent();
-	}
-
-	@Transactional(readOnly = true)
-	public boolean isReviewable(UUID productId) {
-		return productRepository.findByIdAndStatus(productId, ProductStatus.ON_SALE).isPresent();
 	}
 
 	private Product getOnSaleProductOrThrow(UUID productId) {
